+++ src/api/controllers/fighter.controller.js
@@ -1,13 +1,18 @@
 const Fighter = require('../models/fighter.model')
+const Fight = require('../models/fight.model')
+const mongoose = require('mongoose')
 
 // CREATE
 const createFighter = async (req, res) => {
   try {
     const newFighter = new Fighter(req.body)
     const savedFighter = await newFighter.save()
+
     return res.status(201).json(savedFighter)
   } catch (error) {
-    return res.status(400).json(error)
+    return res.status(400).json({
+      message: error.message
+    })
   }
 }
 
@@ -15,9 +20,12 @@ const createFighter = async (req, res) => {
 const getFighters = async (req, res) => {
   try {
     const fighters = await Fighter.find()
+
     return res.status(200).json(fighters)
   } catch (error) {
-    return res.status(400).json(error)
+    return res.status(400).json({
+      message: error.message
+    })
   }
 }
 
@@ -25,10 +33,22 @@ const getFighters = async (req, res) => {
 const updateFighter = async (req, res) => {
   try {
     const { id } = req.params
-    const updated = await Fighter.findByIdAndUpdate(id, req.body, { new: true })
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(400).json({
+        message: 'ID no válido'
+      })
+    }
+
+    const updated = await Fighter.findByIdAndUpdate(id, req.body, {
+      new: true
+    })
+
     return res.status(200).json(updated)
   } catch (error) {
-    return res.status(400).json(error)
+    return res.status(400).json({
+      message: error.message
+    })
   }
 }
 
@@ -36,10 +56,25 @@ const updateFighter = async (req, res) => {
 const deleteFighter = async (req, res) => {
   try {
     const { id } = req.params
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(400).json({
+        message: 'ID no válido'
+      })
+    }
+
+    await Fight.updateMany(
+      { fighters: id },
+      { $pull: { fighters: id } }
+    )
+
     await Fighter.findByIdAndDelete(id)
+
     return res.status(200).json('Fighter eliminado')
   } catch (error) {
-    return res.status(400).json(error)
+    return res.status(400).json({
+      message: error.message
+    })
   }
 }
